+++ blackhole/rendering/Renderer.h
@@ -36,7 +36,7 @@ public:
 	bool loadBlackholeInfluenceRadius(const float blackholeInfluenceRadius) const;
 
 	bool loadLightSpeed(float light_speed) const;
-	bool loadLightStepAmount(uint16_t light_steps) const;
+	bool loadLightStepAmount(uint32_t light_steps) const;
 
 	bool render(char* outputFrame) const;
 };
